# app/main.py
# generated by fastapi-codegen:
#   filename:  ns-provision-openapi.yml
#   timestamp: 2024-01-16T21:00:16+00:00

from __future__ import annotations

from typing import Optional

from fastapi import FastAPI
from fastapi.staticfiles import StaticFiles
from models import ClusterArray, Cluster, NamespaceArray, Provisioning, ProvisioningArray

app = FastAPI(
    title='Namespace provisioning',
    version='1.0',
)

app.mount('/static', StaticFiles(directory='static',html=True))

@app.get('/clusters', response_model=ClusterArray)
def get_clusters() -> ClusterArray:
    return [
        Cluster(value="Portal Gun", text="42.0", csi=True, production=True),
    ]
    pass


@app.get('/namespaces', response_model=NamespaceArray)
def get_namespaces(
    cluster_id: Optional[str] = None, account_id: Optional[str] = None
) -> NamespaceArray:
    pass


@app.get('/provision-namespaces', response_model=ProvisioningArray)
def get_provision_namespaces() -> ProvisioningArray:
    pass


@app.post('/provision-namespaces', response_model=None)
def post_provision_namespaces(body: Provisioning = None) -> None:
    pass